fix: fluxo cupom carrinho deletar

+++ src/main/java/api/loja/lotus/services/CarrinhoService.java
@@ -180,7 +180,7 @@ public class CarrinhoService {
 
         long quantidadeItens = itemCarrinhoRepository.countByCarrinho(carrinho);
 
-        if (quantidadeItens == 0) {
+        if (quantidadeItens == 0 && carrinho.getCupom() == null) {
             carrinhoRepository.delete(carrinho);
             return;
         }
